move network-related classes to their own package
