+++ common/src/main/java/com/abrenoch/hyperiongrabber/common/HyperionScreenEncoder.java
@@ -13,6 +13,8 @@ import android.os.Build;
 import android.support.annotation.RequiresApi;
 import android.util.Log;
 
+import com.abrenoch.hyperiongrabber.common.network.HyperionThread;
+
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.nio.ByteBuffer;
